Fixing forced batch tests to be symlinks and fixing usage of SetForkID.
