+++ test/e2e/shared.go
@@ -5,6 +5,12 @@ import (
 	"context"
 	"fmt"
 	"math/big"
+	"testing"
+	"time"
+
+	"github.com/0xPolygonHermez/zkevm-node/etherman/smartcontracts/polygonzkevm"
+	"github.com/0xPolygonHermez/zkevm-node/test/constants"
+	"github.com/stretchr/testify/require"
 
 	"github.com/0xPolygonHermez/zkevm-node/log"
 	"github.com/0xPolygonHermez/zkevm-node/state"
@@ -20,6 +26,7 @@ const (
 	invalidParamsErrorCode = -32602
 	toAddressHex           = "0x4d5Cf5032B2a844602278b01199ED191A86c93ff"
 	gerFinalityBlocks      = uint64(250)
+	forkID6                = 6
 )
 
 var (
@@ -122,4 +129,99 @@ func logTx(tx *ethTypes.Transaction) {
 	// b, _ := tx.MarshalBinary()
 	//log.Debugf("RLP: ", hex.EncodeToHex(b))
 	log.Debugf("********************")
+}
+
+func sendForcedBatchForVector(t *testing.T, txs []byte, opsman *operations.Manager) (*state.Batch, error) {
+	ctx := context.Background()
+	st := opsman.State()
+	// Connect to ethereum node
+	ethClient, err := ethclient.Dial(operations.DefaultL1NetworkURL)
+	require.NoError(t, err)
+
+	// Create smc client
+	zkEvmAddr := common.HexToAddress(operations.DefaultL1ZkEVMSmartContract)
+	zkEvm, err := polygonzkevm.NewPolygonzkevm(zkEvmAddr, ethClient)
+	require.NoError(t, err)
+
+	auth, err := operations.GetAuth(operations.DefaultSequencerPrivateKey, operations.DefaultL1ChainID)
+	require.NoError(t, err)
+
+	log.Info("Using address: ", auth.From)
+	num, err := zkEvm.LastForceBatch(&bind.CallOpts{Pending: false})
+	require.NoError(t, err)
+	log.Info("Number of forceBatches in the smc: ", num)
+
+	// Get tip
+	tip, err := zkEvm.GetForcedBatchFee(&bind.CallOpts{Pending: false})
+	require.NoError(t, err)
+
+	disallowed, err := zkEvm.IsForcedBatchDisallowed(&bind.CallOpts{Pending: false})
+	require.NoError(t, err)
+	if disallowed {
+		tx, err := zkEvm.ActivateForceBatches(auth)
+		require.NoError(t, err)
+		err = operations.WaitTxToBeMined(ctx, ethClient, tx, operations.DefaultTimeoutTxToBeMined)
+		require.NoError(t, err)
+	}
+
+	// Send forceBatch
+	tx, err := zkEvm.ForceBatch(auth, txs, tip)
+	require.NoError(t, err)
+
+	log.Info("Forced Batch Submit to L1 TxHash: ", tx.Hash())
+	time.Sleep(1 * time.Second)
+
+	err = operations.WaitTxToBeMined(ctx, ethClient, tx, operations.DefaultTimeoutTxToBeMined)
+	require.NoError(t, err)
+
+	currentBlock, err := ethClient.BlockByNumber(ctx, nil)
+	require.NoError(t, err)
+	log.Debug("currentBlock.Time(): ", currentBlock.Time())
+
+	query := ethereum.FilterQuery{
+		FromBlock: currentBlock.Number(),
+		Addresses: []common.Address{zkEvmAddr},
+	}
+	logs, err := ethClient.FilterLogs(ctx, query)
+	require.NoError(t, err)
+
+	var forcedBatch *state.Batch
+	for _, vLog := range logs {
+		if vLog.Topics[0] != constants.ForcedBatchSignatureHash {
+			logs, err = ethClient.FilterLogs(ctx, query)
+			require.NoError(t, err)
+			continue
+		}
+		fb, err := zkEvm.ParseForceBatch(vLog)
+		if err != nil {
+			log.Errorf("failed to parse force batch log event, err: ", err)
+		}
+		log.Debugf("log decoded: %+v", fb)
+		ger := fb.LastGlobalExitRoot
+		log.Info("GlobalExitRoot: ", ger)
+		log.Info("Transactions: ", common.Bytes2Hex(fb.Transactions))
+		fullBlock, err := ethClient.BlockByHash(ctx, vLog.BlockHash)
+		if err != nil {
+			log.Errorf("error getting hashParent. BlockNumber: %d. Error: %v", vLog.BlockNumber, err)
+			return nil, err
+		}
+		log.Info("MinForcedTimestamp: ", fullBlock.Time())
+		forcedBatch, err = st.GetBatchByForcedBatchNum(ctx, fb.ForceBatchNum, nil)
+		for err == state.ErrStateNotSynchronized {
+			time.Sleep(1 * time.Second)
+			forcedBatch, err = st.GetBatchByForcedBatchNum(ctx, fb.ForceBatchNum, nil)
+		}
+		require.NoError(t, err)
+		require.NotNil(t, forcedBatch)
+
+		log.Info("Waiting Forced Batch to be virtualized ...")
+		err = operations.WaitBatchToBeVirtualized(forcedBatch.BatchNumber, 4*time.Minute, st)
+		require.NoError(t, err)
+
+		log.Info("Waiting Forced Batch to be consolidated ...")
+		err = operations.WaitBatchToBeConsolidated(forcedBatch.BatchNumber, 4*time.Minute, st)
+		require.NoError(t, err)
+	}
+
+	return forcedBatch, nil
 }
